Revert "Bug 969215 - Execute |make| should only build specific apps we modified r=me

+++ build/optimize-clean.js
@@ -1,22 +1,41 @@
+/* global exports */
+/* global require */
 'use strict';
 
-/* global exports, require */
-
 var utils = require('./utils');
+var config;
+
+function debug(str) {
+  //dump(' -*- l10n-clean.js: ' + str + '\n');
+}
 
 function execute(options) {
   var gaia = utils.gaia.getInstance(options);
+  config = options;
+  debug('Begin');
 
-  gaia.rebuildWebapps.forEach(function(webapp) {
-    let re = new RegExp('\\.html\\.' + options.GAIA_DEFAULT_LOCALE + '$');
+  gaia.webapps.forEach(function(webapp) {
+    // if BUILD_APP_NAME isn't `*`, we only accept one webapp
+    if (config.BUILD_APP_NAME != '*' &&
+      webapp.sourceDirectoryName != config.BUILD_APP_NAME) {
+      return;
+    }
+
+    debug(webapp.sourceDirectoryName);
+
+    let re = new RegExp('\\.html\\.' + config.GAIA_DEFAULT_LOCALE + '$');
     let files = utils.ls(webapp.buildDirectoryFile, true);
     files.forEach(function(file) {
-      if (re.test(file.leafName) ||
-        file.leafName.indexOf(gaia.aggregatePrefix) === 0) {
+      if (
+        re.test(file.leafName) ||
+        file.leafName.indexOf(gaia.aggregatePrefix) === 0
+      ) {
         file.remove(false);
       }
     });
   });
+
+  debug('End');
 }
 
 exports.execute = execute;
